+++ extensions/amp-user-notification/amp-user-notification.md
@@ -29,10 +29,7 @@ limitations under the License.
       <div>
         <code>&lt;script async custom-element="amp-user-notification" src="https://cdn.ampproject.org/v0/amp-user-notification-0.1.js">&lt;/script></code>
       </div>
-      <div>
-        <code>&lt;script async custom-element="amp-analytics" src="https://cdn.ampproject.org/v0/amp-analytics-0.1.js">&lt;/script></code>
-      </div>
-      <small>Notice that both "amp-user-notification" and "amp-analytics" scripts are required.</small>
+      <small>Notice that  "amp-user-notification" script is required.</small>
     </td>
   </tr>
   <tr>
@@ -72,7 +69,7 @@ adding a unique value to the `GET` url like a timestamp as a query string field)
 
 When multiple `amp-user-notification` elements are on a page, only one is shown
 at a single time (Once one is dismissed the next one is shown).
-The order of the notifications being shown is currently not deterministic. 
+The order of the notifications being shown is currently not deterministic.
 
 Example:
 
@@ -97,7 +94,7 @@ on the `data-show-if-href` attribute (see [#1228](https://github.com/ampproject/
 
 As a best practice to not let the browser cache the GET response values, you should add
 a [`TIMESTAMP` url replacement](https://github.com/ampproject/amphtml/blob/master/spec/amp-var-substitutions.md) value to the `data-show-if-href` attribute value.
-You can add it as a query string field (e.g., 
+You can add it as a query string field (e.g.,
 `data-show-if-href="https://foo.com/api/show-api?timestamp=TIMESTAMP"`).
 
 If the `data-show-if-href` attribute is not specified, AMP will only check if the notification with the specified ID has been "dismissed" by the user locally. If not, the notification will be shown.
@@ -193,7 +190,7 @@ This notification should ALWAYS show on every page visit.
 
 A notification is shown when:
 
-1. There's no record locally that the user has dismissed the notification with the 
+1. There's no record locally that the user has dismissed the notification with the
 specified ID.
 2. When specified, `data-show-if-href` endpoint returns `{ "showNotification": true }`.
